Move constants out of pundit.rb into conventional location
